requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
cycler==0.11.0
environs==9.3.5
fonttools==4.28.2
imageio==2.12.0
install==1.3.4
kiwisolver==1.3.2
marshmallow==3.14.1
matplotlib==3.5.0
networkx==2.6.3
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
scikit-image==0.18.3
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
tifffile==2021.11.2
tomli==1.2.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
